+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All Notable changes to `Csv` will be documented in this file
 
 - `Statement::select`
 - `TabularDataReader::getRecordsAsObject`
+- `TabularDataReader::chunkBy`
 
 ### Deprecated